Apply some migrations

+++ Homeworks/06.Entity Framework Core/05.EntityRelations/P03_FootballBetting.Data.Models/PlayerStatistic.cs
@@ -12,6 +12,6 @@
 
         public int Assists { get; set; }
 
-        public double MinutesPlayed { get; set; }
+        public int MinutesPlayed { get; set; }
     }
 }
